+++ app/src/main/java/com/TheGingerMan/spaceobjects/Vector.java
@@ -2,15 +2,15 @@ package com.TheGingerMan.spaceobjects;
 
 public class Vector {
 
-    float x, y, z;
+    double x, y, z;
 
-    public Vector(float x, float y, float z) {
+    public Vector(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public Vector(float x, float y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
         z=0;
